refactor: docker-compose upstream transition
Prior to this, the CLI was relying on a fork of PDMLab/docker-compose due to the latter lacking support for Docker Compose v2.
Upstream has finally integrated (somewhat acceptable) support for v2.

This change reintroduces the upstream library in hopes of reducing code duplication and avoiding divergence issues down the line.